requirements.txt
flask
flask_cors
jwt
azure-storage-blob 
azure-identity
pytest 
Flask-Testing 
pytest-mock